cmake/libuv:  edits so cmake would build
cmake needed a dependency on libuv to build.

libuv can't be built --with-check on older Darwins, because one of the tests
relies on a newer version of a system header.  (On Leopard, and probably
several versions after it, there is no `birthtime` field in a `struct stat`.)
This still needs a patch.

Added --universal to libuv’s `test`.

Tweaked the ./configure option for quietness vs. verbosity so that Tigerbrew’s
`ARGV.verbose?` setting controls it.